+++ frontend/src/config/index.ts
@@ -2,3 +2,4 @@
  * Config exports
  */
 export * from './uiConfig'
+export * from './saveConfig'